Add port to the Nmap command used for Problem 4
I don't entirely remember exactly which flags were used for the commands
in Problem 4 so the description there may be slightly incorrect,
although on the whole it doesn't make a significant difference. I
believe that the logs from Nmap were generated with `-p 80`, and in any
case adding that flag wouldn't change the logs we're seeing (just make
the scan faster), so I've added that flag there.