Start adding container image for Lab 1.
Lab 1 was a little unusual compared to the later labs since it was
created entirely with Ansible. Since the main goal of this repository is
to do everything with Docker, I need to rewrite a lot of the lab from
scratch in order to get it working correctly.

This first commit adds a base Dockerfile that installs OpenSSH, adds
each of the required users, and sets up Problems 1 and 2. Over time I'll
progressively add more of the problems until the lab includes all of the
problems of the original.